[TASK 1] Bugs fixes made

+++ task1/Merge/merge.h
@@ -3,6 +3,6 @@
 
 #include "../Array/array_and_size.h"
 
-void merge_all_parts(ArrayAndSize *array_size, int *temp, long *bounds, int parts_count);
+void merge_all_parts(ArrayAndSize *array_size, int *temp, long *bounds, long parts_count);
 
 #endif //MERGE
